backEnd/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0
redis==5.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
geopy==2.4.1
pytz==2024.1
